+++ src/sql/run/run.py
@@ -33,6 +33,13 @@ def run_statements(conn, sql, config, parameters=None):
         return "Connected: %s" % conn.name
 
     for statement in sqlparse.split(sql):
+        # strip all comments from sql
+        statement = sqlparse.format(statement, strip_comments=True)
+        # trailing comment after semicolon can be confused as its own statement,
+        # so we ignore it here.
+        if not statement:
+            continue
+
         first_word = sql.strip().split()[0].lower()
 
         if first_word == "begin":