+++ src/tests/test_magic.py
@@ -2248,3 +2248,87 @@ INSERT INTO languages VALUES ('Python', 1), ('Java', 0), ('OCaml', 2)"""
 )
 def test_get_query_type(query, query_type):
     assert get_query_type(query) == query_type
+
+
+@pytest.mark.parametrize(
+    "query, expected",
+    [
+        (
+            """%%sql
+SELECT 1""",
+            1,
+        ),
+        (
+            """%%sql
+SELECT 1 -- comment""",
+            1,
+        ),
+        (
+            """%%sql
+SELECT 1
+-- comment""",
+            1,
+        ),
+        (
+            """%%sql
+SELECT 1; -- comment""",
+            1,
+        ),
+        (
+            """%%sql
+SELECT 1;
+-- comment""",
+            1,
+        ),
+        (
+            """%%sql
+-- comment before
+SELECT 1;""",
+            1,
+        ),
+        (
+            """%%sql
+-- comment before
+SELECT 1;
+-- comment after""",
+            1,
+        ),
+        (
+            """%%sql
+SELECT 1; -- comment
+SELECT 2""",
+            2,
+        ),
+        (
+            """%%sql
+SELECT 1; -- comment
+SELECT 2;""",
+            2,
+        ),
+        (
+            """%%sql
+SELECT 1;
+-- comment
+SELECT 2;""",
+            2,
+        ),
+        (
+            """%%sql
+SELECT 1;
+-- comment before
+SELECT 2;
+-- comment after""",
+            2,
+        ),
+        (
+            """%%sql
+SELECT 1; -- comment before
+SELECT 2;
+-- comment after""",
+            2,
+        ),
+    ],
+)
+def test_query_comment_after_semicolon(ip, query, expected):
+    result = ip.run_cell(query).result
+    assert list(result.dict().values())[-1][0] == expected